pyproject.toml
[project]
name = "phf-search"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "colabdesign @ git+https://github.com/sokrypton/ColabDesign.git@v1.1.1",
    "biopython>=1.80",
    "numpy",
    "scipy",
]

[project.optional-dependencies]
cuda = [
    "jax[cuda12]",
]
cpu = [
    "jax",
]
